+++ src/main/java/gg/projecteden/nexus/features/minigames/models/matchdata/PixelDropMatchData.java
@@ -10,7 +10,6 @@ import gg.projecteden.nexus.features.minigames.models.annotations.MatchDataFor;
 import gg.projecteden.nexus.features.minigames.models.arenas.PixelDropArena;
 import gg.projecteden.nexus.framework.exceptions.NexusException;
 import gg.projecteden.nexus.utils.ActionBarUtils;
-import gg.projecteden.nexus.utils.Nullables;
 import gg.projecteden.nexus.utils.RandomUtils;
 import gg.projecteden.nexus.utils.StringUtils;
 import gg.projecteden.utils.TimeUtils.TickTime;
@@ -49,7 +48,7 @@ public class PixelDropMatchData extends MatchData {
 	private int wordTaskId;
 	private int currentRound;
 	private long roundStart;
-	private int timeLeft;
+	private long timeLeft;
 	private int roundCountdownId;
 	private boolean roundOver;
